+++ src/modules/mavlink/mavlink_messages.cpp
@@ -121,6 +121,7 @@ using matrix::wrap_2pi;
 #include "streams/HIL_ACTUATOR_CONTROLS.hpp"
 #include "streams/HIL_STATE_QUATERNION.hpp"
 #include "streams/HOME_POSITION.hpp"
+#include "streams/LOCAL_POSITION_NED.hpp"
 #include "streams/MANUAL_CONTROL.hpp"
 #include "streams/MOUNT_ORIENTATION.hpp"
 #include "streams/NAV_CONTROLLER_OUTPUT.hpp"
@@ -2720,74 +2721,6 @@ protected:
 	}
 };
 
-class MavlinkStreamLocalPositionNED : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamLocalPositionNED::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "LOCAL_POSITION_NED";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_LOCAL_POSITION_NED;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamLocalPositionNED(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return _lpos_sub.advertised() ? MAVLINK_MSG_ID_LOCAL_POSITION_NED_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES : 0;
-	}
-
-private:
-	uORB::Subscription _lpos_sub{ORB_ID(vehicle_local_position)};
-
-	/* do not allow top copying this class */
-	MavlinkStreamLocalPositionNED(MavlinkStreamLocalPositionNED &) = delete;
-	MavlinkStreamLocalPositionNED &operator = (const MavlinkStreamLocalPositionNED &) = delete;
-
-protected:
-	explicit MavlinkStreamLocalPositionNED(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		vehicle_local_position_s lpos;
-
-		if (_lpos_sub.update(&lpos)) {
-			mavlink_local_position_ned_t msg{};
-
-			msg.time_boot_ms = lpos.timestamp / 1000;
-			msg.x = lpos.x;
-			msg.y = lpos.y;
-			msg.z = lpos.z;
-			msg.vx = lpos.vx;
-			msg.vy = lpos.vy;
-			msg.vz = lpos.vz;
-
-			mavlink_msg_local_position_ned_send_struct(_mavlink->get_channel(), &msg);
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
 class MavlinkStreamCameraCapture : public MavlinkStream
 {
 public:
@@ -2895,7 +2828,9 @@ static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamSystemTime>(),
 	create_stream_list_item<MavlinkStreamTimesync>(),
 	create_stream_list_item<MavlinkStreamGlobalPositionInt>(),
+#if defined(LOCAL_POSITION_NED_HPP)
 	create_stream_list_item<MavlinkStreamLocalPositionNED>(),
+#endif // LOCAL_POSITION_NED_HPP
 	create_stream_list_item<MavlinkStreamOdometry>(),
 #if defined(ESTIMATOR_STATUS_HPP)
 	create_stream_list_item<MavlinkStreamEstimatorStatus>(),
